// packages/twenty-apps/public/zoominfo/src/types/zoominfo-contact-match-input.ts
export type ZoomInfoContactMatchInput = {
  personId?: number;
  fullName?: string;
  firstName?: string;
  lastName?: string;
  emailAddress?: string;
  phone?: string;
  jobTitle?: string;
  companyId?: number;
  companyName?: string;
  externalURL?: string;
  contactAccuracyScoreMin?: number;
};